pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "abprop"
version = "0.1.0"
description = "Sequence-only antibody property modeling toolkit."
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "AbProp Team", email = "ml@abprop.local"}
]
keywords = ["antibody", "protein", "machine-learning", "pytorch"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "numpy>=1.24,<2",
    "pandas>=2.0",
    "pyarrow>=12.0",
    "torch>=2.1,<2.2",
    "pyyaml>=6.0",
    "tqdm>=4.66",
    "matplotlib>=3.7",
    "requests>=2.31",
]

[project.scripts]
abprop-etl = "abprop.commands.etl:main"
abprop-train = "abprop.commands.train:main"
abprop-eval = "abprop.commands.eval:main"
abprop-launch = "abprop.commands.launch:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "pytest-cov>=4.1",
    "ruff>=0.2.0",
    "mypy>=1.8",
]
serve = [
    "fastapi>=0.110",
    "uvicorn>=0.23",
]
bench = [
    "mlflow>=2.12",
    "seaborn>=0.13",
    "scipy>=1.10,<1.12",
    "scikit-learn>=1.2,<1.4",
]
viz = [
    "plotly>=5.18",
    "scikit-learn>=1.2,<1.4",
    "umap-learn>=0.5.5",
]
dashboard = [
    "streamlit>=1.31",
    "plotly>=5.18",
]

[project.urls]
Homepage = "https://example.com/abprop"
Documentation = "https://example.com/abprop/docs"
Source = "https://example.com/abprop/source"

[tool.setuptools.packages.find]
where = ["src"]
